v3-Move ConfigurableAreas to Own Section WithConfigurableAreas (#1510)
* Move ConfigurableAreas to Own Trait

---------
