+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to `laravel-livewire-tables` will be documented in this file
 
+## UNRELEASED
+- Split ConfigurableArea into own Trait/Config/Helper Files
+
 ## [v3.1.0] - 2023-10-31
 - Restore wire:confirm for Bulk Actions
 - Stable Release of 3.1.0 Stable for Livewire v3 Support
