+++ src/Traits/ComponentUtilities.php
@@ -60,20 +60,6 @@ trait ComponentUtilities
 
     protected array $additionalSelects = [];
 
-    protected bool $hideConfigurableAreasWhenReorderingStatus = true;
-
-    protected array $configurableAreas = [
-        'before-tools' => null,
-        'toolbar-left-start' => null,
-        'toolbar-left-end' => null,
-        'toolbar-right-start' => null,
-        'toolbar-right-end' => null,
-        'before-toolbar' => null,
-        'after-toolbar' => null,
-        'before-pagination' => null,
-        'after-pagination' => null,
-    ];
-
     // Sets the Theme If Not Already Set
     public function mountComponentUtilities(): void
     {
