+++ src/Traits/Configuration/ComponentConfiguration.php
@@ -210,37 +210,6 @@ trait ComponentConfiguration
         return $this;
     }
 
-    /**
-     * @param  array<mixed>  $areas
-     */
-    public function setConfigurableAreas(array $areas): self
-    {
-        $this->configurableAreas = $areas;
-
-        return $this;
-    }
-
-    public function setHideConfigurableAreasWhenReorderingStatus(bool $status): self
-    {
-        $this->hideConfigurableAreasWhenReorderingStatus = $status;
-
-        return $this;
-    }
-
-    public function setHideConfigurableAreasWhenReorderingEnabled(): self
-    {
-        $this->setHideConfigurableAreasWhenReorderingStatus(true);
-
-        return $this;
-    }
-
-    public function setHideConfigurableAreasWhenReorderingDisabled(): self
-    {
-        $this->setHideConfigurableAreasWhenReorderingStatus(false);
-
-        return $this;
-    }
-
     public function setDataTableFingerprint(string $dataTableFingerprint): self
     {
         $this->dataTableFingerprint = $dataTableFingerprint;
